Restructure visibility metrics
